# Implement `UnitBase.components()`