--- a/docs/changes/units/16783.feature.rst
+++ b/docs/changes/units/16783.feature.rst
@@ -1,0 +1,2 @@
+``UnitBase`` instances now have a ``components()`` method, which enables
+convenient looping over the bases and powers of the unit.