### Description

The new method provides a convenient way of looping over the bases and powers of a unit, which previously required accessing both its `bases` and `powers` properties. In some cases this avoids having to store the unit in a variable so that the two properties could be called separately. It also enables [specifying `zip()` with `strict=True`](https://docs.astral.sh/ruff/rules/zip-without-explicit-strict/) once in the definition of `components` instead of having to write out `strict` every time `bases` and `powers` are zipped together.

- [x] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.