@eerovaher 2024-07-29
If `components` would be a `dict[UnitBase, UnitPower]` then it might contain a key that is the unit itself, but that means the unit must already be hashable, so computing the hash could not involve `components` at all. This means that not all calls of `zip(unit.bases, unit.powers)` could be replaced with `unit.components.items()`. I would much prefer an implementation of `components` that could always be used as a replacement.

`components` could return a `tuple[tuple[UnitBase, UnitPower], ...]` because that does not involve any hashing. If there are performance concerns then we could use `@cached_property`, and if there are concerns about API clarity then we could make `components` private.