@mhvk 2024-07-29
Seems unlikely this PR would have *this* big effect, and the error suggests it is just a fluke, but since it might be lreated,  the benchmark failure is due to:
```
| Change   | Before [377b6ca5]    | After [b8ad60e3]    |   Ratio | Benchmark (Parameter)                               |
|----------|----------------------|---------------------|---------|-----------------------------------------------------|
| -        | 357±30μs             | 228±4μs             |    0.64 | units.TimeQuantityOpLargeArray.time_quantity_square |
| Change   | Before [377b6ca5]    | After [b8ad60e3]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| +        | 41.7±0.3μs           | 127±80μs            |    3.04 | units.time_quantity_times_unit |
```