@mhvk 2024-07-29
@eerovaher - yes, I remember the problem with a `dict` - but the idea had elegance... 

Here, I see an increase in complexity for little gain - we have `.bases` and `.powers` and those are clear, consistent with the inputs to `CompositeUnit`, and we `zip` them together in a few places, which to me does not really call out for new API (even internal only).