@nstarman 2024-07-30
I can take a closer look at this some time over this week.
My first thought is that I really like this general idea. A composite unit (where for consistency here all units have the same API) really is just an enumeration of the base units raised to some power.  Neither the bases nor the powers really have that much meaning individually. So to me it makes the most sense to keep them packaged together.