@mhvk 2024-07-30
Do agree that starting from scratch one might design things differently, and always good to think things through.

Fundamentally, any composite unit has a scale multiplied by a series of components, each of which is a base unit raised to some power. Given that, a mapping is not necessarily the most logical (even if I liked it before...) - a `NamedTuple` with `base` and `power` might work just as well. Of course, the latter is much closer to what @eerovaher has here.

But for astropy the API has already been in place for a long time. I remain feeling it is not worth extending it... Indeed, I'm not sure in actual use, much access to the power and bases is needed at all.