@mhvk 2024-07-30
A different thought related to actual use: I recall someone in gammapy (@cdeil?) mentioning that he always uses units as strings since something like "km/s/Mpc" is clearer than `u.km/u.s/u.Mpc`, and scanning of units is fast enough anyway. I always thought he had a point, even if I found scanning rather slow... But that is where a little cache in `u.Unit` would immediately make a big difference: if it had a cache, `u.Unit("km/s/Mpc")` would in fact be much faster than `u.km/u.s/u.Mpc`! (To make you typing types happy, we'd perhaps have to suggest something like `from astropy.units import Unit as U` and use `U("<favourite-unit>")` throughout -- unless we can convince python to introduce a new string type for it which I vaguely recall once seeing suggested ... ... hmm, cannot find it, but an [in-the-works PEP](https://github.com/jimbaker/tagstr/blob/main/pep.rst) would enable it - noted that in https://github.com/jimbaker/tagstr/issues/51