@eerovaher 2024-09-30
I had encountered quite a lot of code in `units` that iterates over both the unit bases and powers (as evidenced by the number of merge conflicts this small patch has accumulated). That left me with an impression that a separate method for such iteration would be worthwhile. When I was working on this I realized that this iteration is actually not very common, but I had already put effort into this so it seemed reasonable to open a pull request anyways. However, by now it's clear that the proposed method would not be used frequently enough to justify its existence.